llm_integration/requirements.txt
openai>=1.30.0
anthropic>=0.25.0
google-generativeai>=0.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
